Delete and update UI

+++ app/src/main/java/com/company/ourfinances/view/viewholder/FinanceRecordViewHolder.kt
@@ -1,5 +1,7 @@
 package com.company.ourfinances.view.viewholder
 
+import android.app.AlertDialog
+import android.content.DialogInterface
 import androidx.lifecycle.LifecycleOwner
 import androidx.recyclerview.widget.RecyclerView
 import com.company.ourfinances.databinding.FinaceRecordViewItemBinding
@@ -23,7 +25,18 @@ class FinanceRecordViewHolder(
         //bind.textTagRecord.text = financeRecordEntity
 
         bind.buttonDeleteView.setOnClickListener {
-            listener.onDelete(financeRecordEntity.recordId)
+
+            AlertDialog.Builder(itemView.context)
+                .setTitle("Remoção de Registro")
+                .setMessage("Tem certeza que deseja excluir o registro? Isso não pode ser revertido.")
+                .setPositiveButton("Sim") { dialog, which ->
+                    listener.onDelete(financeRecordEntity.recordId)
+                }
+                .setNegativeButton("Não", null)
+                .setNeutralButton("Cancelar", null)
+                .create()
+                .show()
+
         }
     }
